move backend organization to src
