+++ backend/tsconfig.json
@@ -1,15 +1,20 @@
 {
   "compilerOptions": {
+    "target": "ES2020",
     "module": "NodeNext",
     "moduleResolution": "NodeNext",
-    "target": "es2020",
     "esModuleInterop": true,
-    "allowJs": true,
+    "forceConsistentCasingInFileNames": true,
     "strict": true,
-    "skipLibCheck": true
+    "skipLibCheck": true,
+    "allowJs": true,
+    "resolveJsonModule": true,
+    "moduleDetection": "force",
+    "allowImportingTsExtensions": true,
+    "declaration": true,
+    "emitDeclarationOnly": true,
+    "isolatedModules": true
   },
-  "ts-node": {
-    "esm": true,
-    "experimentalSpecifiers": true
-  }
+  "include": ["src/**/*"],
+  "exclude": ["node_modules"]
 }
